kola: simplify tests on non-QEMU
We only need to worry about basic-qemu-scenarios and running reprovision
tests separately when the platform is QEMU (the default). Loosely detect
if the target platform is something other than QEMU by checking to see
if the user provided platformArgs and if we are running against non-QEMU
then just run all the tests up front and not in separate runs.