{kola,kolaTestIso}: workaround issue with accessDenied from dir()
When using `dir()` in the kola and kolaTestIso jobs the cd is happening
on the controller (not in the COSA pod) so if the directory
isn't under the env.WORKSPACE directory then we get a permission denied
error.

Let's go back to using a cd ${cosaDir} for now to workaround this while
we find better solutions. This cd ${cosaDir} will have no effect on
multi-arch where we'll be operating in a remote session anyway.

more tests