{kola,kolaTestIso}: remove Arch from stage run titles by default
Including the arch everywhere is a bit overboard. Where it's really
needed is in our pipeline bump-lockfile job where we run tests from
all architectures at once. Everywhere else it's not really welcome.

Let's change the model a bit to include the marker given by the client
in the stage titles. Otherwise we won't add anything to them.